# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Chase Bank Online Banking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  username: string;
  password: string;
  rememberMe?: boolean;
}

export interface UserProfile {
  id: number;
  fullName: string;
  username: string;
  email?: string;
  phone?: string;
  address?: string;
  accountNumber: string;
  accountType: string;
  accountStatus: string;
  availableBalance: number;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user?: UserProfile;
  isAdmin: boolean;
}

export interface UpdateProfileRequest {
  fullName?: string;
  email?: string;
  phone?: string;
  address?: string;
}

export interface Transaction {
  id: number;
  transactionReference: string;
  beneficiaryName?: string;
  bankName?: string;
  accountNumber?: string;
  routingNumber?: string;
  amount: number;
  currency: string;
  debit?: number;
  credit?: number;
  balanceAfter?: number;
  description?: string;
  status: string;
  createdAt: string;
}

export interface DashboardSummary {
  user: UserProfile;
  currentBalance: number;
  monthlyDeposits: number;
  monthlyWithdrawals: number;
  totalTransactions: number;
  recentTransactions: Transaction[];
}

export interface TransactionListResponse {
  transactions: Transaction[];
  total: number;
  page: number;
  limit: number;
}

export interface TransferRequest {
  beneficiaryName: string;
  bankName: string;
  accountNumber: string;
  routingNumber: string;
  country: string;
  amount: number;
  reference?: string;
}

export interface TransferResponse {
  id: number;
  transactionReference: string;
  beneficiaryName: string;
  bankName: string;
  accountNumber: string;
  routingNumber: string;
  amount: number;
  currency: string;
  status: string;
  createdAt: string;
  estimatedCompletion: string;
}

export interface TransferListResponse {
  transfers: TransferResponse[];
  total: number;
}

export interface TransferReceiptResponse {
  transfer: TransferResponse;
  senderName: string;
  senderAccountNumber: string;
}

export interface AdminStats {
  totalUsers: number;
  activeUsers: number;
  suspendedUsers: number;
  totalDeposits: number;
  totalWithdrawals: number;
  pendingTransfers: number;
  completedTransfers: number;
}

export interface AdminUser {
  id: number;
  fullName: string;
  username: string;
  email?: string;
  phone?: string;
  address?: string;
  accountNumber: string;
  accountType: string;
  accountStatus: string;
  availableBalance: number;
  createdAt?: string;
}

export interface AdminUserListResponse {
  users: AdminUser[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateUserRequest {
  fullName: string;
  username: string;
  password: string;
  email?: string;
  phone?: string;
  address?: string;
  accountNumber?: string;
  accountType: string;
  accountStatus: string;
  availableBalance: number;
}

export interface UpdateUserRequest {
  fullName?: string;
  email?: string;
  phone?: string;
  address?: string;
  accountType?: string;
  accountStatus?: string;
}

export type UpdateStatusRequestStatus = typeof UpdateStatusRequestStatus[keyof typeof UpdateStatusRequestStatus];


export const UpdateStatusRequestStatus = {
  Active: 'Active',
  Suspended: 'Suspended',
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
}

export type UpdateBalanceRequestAction = typeof UpdateBalanceRequestAction[keyof typeof UpdateBalanceRequestAction];


export const UpdateBalanceRequestAction = {
  set: 'set',
  increase: 'increase',
  decrease: 'decrease',
} as const;

export interface UpdateBalanceRequest {
  action: UpdateBalanceRequestAction;
  amount: number;
}

export type AddTransactionRequestType = typeof AddTransactionRequestType[keyof typeof AddTransactionRequestType];


export const AddTransactionRequestType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  debit: 'debit',
  credit: 'credit',
} as const;

export interface AddTransactionRequest {
  type: AddTransactionRequestType;
  amount: number;
  description?: string;
  transactionDate?: string;
}

export interface AdminTransfer {
  id: number;
  transactionReference: string;
  userId: number;
  userName?: string;
  userAccountNumber?: string;
  beneficiaryName: string;
  bankName: string;
  accountNumber: string;
  routingNumber?: string;
  amount: number;
  currency: string;
  status: string;
  reference?: string;
  createdAt: string;
}

export interface AdminTransferListResponse {
  transfers: AdminTransfer[];
  total: number;
}

export type UpdateTransferStatusRequestStatus = typeof UpdateTransferStatusRequestStatus[keyof typeof UpdateTransferStatusRequestStatus];


export const UpdateTransferStatusRequestStatus = {
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export interface UpdateTransferStatusRequest {
  status: UpdateTransferStatusRequestStatus;
}

export type GetTransactionsParams = {
page?: number;
limit?: number;
status?: GetTransactionsStatus;
};

export type GetTransactionsStatus = typeof GetTransactionsStatus[keyof typeof GetTransactionsStatus];


export const GetTransactionsStatus = {
  Processing: 'Processing',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export type GetAdminUsersParams = {
page?: number;
limit?: number;
search?: string;
};

export type GetAdminTransfersParams = {
status?: GetAdminTransfersStatus;
search?: string;
};

export type GetAdminTransfersStatus = typeof GetAdminTransfersStatus[keyof typeof GetAdminTransfersStatus];


export const GetAdminTransfersStatus = {
  Processing: 'Processing',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;