# lib/api-zod/src/generated/types/addTransactionRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Chase Bank Online Banking API
 * OpenAPI spec version: 0.1.0
 */
import type { AddTransactionRequestType } from './addTransactionRequestType';

export interface AddTransactionRequest {
  type: AddTransactionRequestType;
  amount: number;
  description?: string;
}